MQTT/requirements.txt
paho-mqtt==2.1.0
tomli==2.0.1